+++ lib/smee/security/xmlsec1.ex
@@ -14,7 +14,6 @@ defmodule Smee.Security.Xmlsec1 do
     cert_file = SigningCertificate.prepare_file!(metadata)
 
     command = build_command(metadata, cert_file)
-
     case Rambo.run("xmlsec1", command, in: metadata.data, log: false) do
       {:ok, %Rambo{status: 0, out: _out}} -> Map.merge(metadata, %{verified: true})
       {:error, %Rambo{status: status, err: err}} -> raise(parse_error(status, err))